napravljena POST ruta za nalaze

+++ PatientService/src/main/java/ba/unsa/etf/nwt/PatientService/repositories/TestRepository.java
@@ -3,5 +3,5 @@ package ba.unsa.etf.nwt.PatientService.repositories;
 import ba.unsa.etf.nwt.PatientService.model.Test;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestRepository extends JpaRepository<Test, Integer> {
+public interface TestRepository extends JpaRepository<Test, Long> {
 }
