+++ PatientService/src/main/java/ba/unsa/etf/nwt/PatientService/repositories/TestTypeRepository.java
@@ -3,5 +3,5 @@ package ba.unsa.etf.nwt.PatientService.repositories;
 import ba.unsa.etf.nwt.PatientService.model.TestType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestTypeRepository extends JpaRepository<TestType, Integer> {
+public interface TestTypeRepository extends JpaRepository<TestType, Long> {
 }
